refactor(pub-sub): simplify pub-sub mechanism
